finished project

+++ scripts/script.js
@@ -1,8 +1,10 @@
+// wird beim laden der Seite ausgeführt
 function init() {
     renderMeals();
     renderToCard();
 }
 
+//  rendert den food_Container mit Daten aus dem Array
 function renderMeals() {
     const contentRef = document.getElementById("food_Container");
     contentRef.innerHTML = "";
@@ -11,6 +13,7 @@ function renderMeals() {
     }
 }
 
+//  Fügt das ausgewählte Produkt dem Warenkorb hinzu
 function addToCard(i) {
     let cardContRef = document.getElementById("basket");
     if (document.getElementById(`${meals[i].name}`)) {
@@ -22,6 +25,7 @@ function addToCard(i) {
     }
 }
 
+//  Löscht das ausgewählte Produkt aus dem Warenkorb (LocalStorage) raus
 function removeFromCard(i) {
     let deleteRef = document.getElementById(`${meals[i].name}`);
     if (deleteRef) {
@@ -32,71 +36,89 @@ function removeFromCard(i) {
     }
 }
 
-function calculatePrice(w, i) {
-    let localRef = JSON.parse(localStorage.getItem(i));
-    let multiWert = parseInt(localRef[1]);
-    let priceWert = parseFloat(localRef[0]);
+function loadLocalData(i) {
+    return JSON.parse(localStorage.getItem(i));
+}
 
-    switchFunction(w, i, multiWert, priceWert);
+function updateLocalData(i, newPrice, newMultiplier) {
+    saveToLocalStorage(i, newPrice.toFixed(2), newMultiplier);
+}
+
+function handleDecrease(i, priceWert, multiWert) {
+    if (multiWert > 1) {
+        updateLocalData(i, priceWert - meals[i].price, multiWert - 1);
+    } else {
+        removeFromCard(i);
+    }
     calculateTotalPrice();
     renderToCard();
-    // switch (w) {
-    //     case "-":
-    //         if (multiWert > 1) {
-    //             saveToLocalStorage(
-    //                 i,
-    //                 (priceWert - meals[i].price).toFixed(2),
-    //                 multiWert - 1
-    //             );
-    //             calculateTotalPrice();
-    //             renderToCard();
-    //         } else {
-    //             removeFromCard(i);
-    //             calculateTotalPrice();
-    //             renderToCard();
-    //         }
+}
 
-    //         break;
+function handleIncrease(i, priceWert, multiWert) {
+    updateLocalData(i, priceWert + meals[i].price, multiWert + 1);
+    calculateTotalPrice();
+    renderToCard();
+}
 
-    //     case "+":
-    //         saveToLocalStorage(
-    //             i,
-    //             (priceWert + meals[i].price).toFixed(2),
-    //             multiWert + 1
-    //         );
-    //         calculateTotalPrice();
-    //         renderToCard();
-    //         break;
+function resetMultiplier(i) {
+    let multiplierTemplateRef = document.getElementById(`multiplier${i}`);
+    multiplierTemplateRef.innerHTML = 1;
+    calculateTotalPrice();
+    renderToCard();
+}
 
-    //     default:
-    //         multiplierTemplateRef.innerHTML = "";
-    //         multiplierTemplateRef.innerHTML = 1;
-    //         calculateTotalPrice();
-    //         renderToCard();
-    //         break;
-    // }
+function calculatePrice(w, i) {
+    let localRef = loadLocalData(i);
+    let priceWert = parseFloat(localRef[0]);
+    let multiWert = parseInt(localRef[1]);
+
+    switch (w) {
+        case "-":
+            handleDecrease(i, priceWert, multiWert);
+            break;
+        case "+":
+            handleIncrease(i, priceWert, multiWert);
+            break;
+        default:
+            resetMultiplier(i);
+            break;
+    }
+}
+
+function getLocalStorageData() {
+    let localRef = [];
+    for (let z = 0; z < localStorage.length; z++) {
+        let localKeyRef = localStorage.key(z);
+        localRef.push(JSON.parse(localStorage.getItem(localKeyRef)));
+    }
+    return localRef;
+}
+
+function calculateTotalFromData(localRef) {
+    let totalPrice = 0;
+    for (let i = 0; i < localRef.length; i++) {
+        totalPrice += parseFloat(localRef[i][0]);
+    }
+    return totalPrice;
+}
+
+function updateTotalPriceUI(totalPrice) {
+    let totalPriceRef = document.getElementById("totalPrice");
+    totalPriceRef.innerHTML =
+        "Gesamtpreis: " +
+        parseFloat(totalPrice).toFixed(2) +
+        "&nbsp;€&nbsp;<button id='toPay' class='to_PayBtn' onclick='toPay()'>Bestellen</button>";
+
+    let toPayRef = document.getElementById("toPayFinal");
+    if (toPayRef) {
+        toPayRef.innerHTML = "";
+    }
 }
 
 function calculateTotalPrice() {
-    let totalPriceRef = document.getElementById("totalPrice");
-    // let classMealPriceRef = document.getElementsByClassName("priceMealW");
-    // let totalPrice = 0;
-    // let localRef = [];
-    // for (let z = 0; z < localStorage.length; z++) {
-    //     let localKeyRef = localStorage.key(z);
-    //     localRef.push(JSON.parse(localStorage.getItem(localKeyRef)));
-    // }
-    // for (let i = 0; i < localRef.length; i++) {
-    //     totalPrice += parseFloat(localRef[i][0]);
-    // }
-    // totalPriceRef.innerHTML = "";
-    // totalPriceRef.innerHTML =
-    //     "Gesamtpreis: " +
-    //     parseFloat(totalPrice).toFixed(2) +
-    //     "&nbsp;€&nbsp;<button id='toPay' class='to_PayBtn' onclick='toPay()'>Bestellen</button>";
-    // let toPayRef = document.getElementById("toPayFinal");
-    // toPayRef.innerHTML = "";
-    totalPriceTemplate(totalPriceRef);
+    let localRef = getLocalStorageData();
+    let totalPrice = calculateTotalFromData(localRef);
+    updateTotalPriceUI(totalPrice);
 }
 
 function saveToLocalStorage(i, price, multiplier) {
@@ -119,10 +141,12 @@ function renderToCard() {
 function toPay() {
     let totalPriceRef = document.getElementById("totalPrice");
     let toPayRef = document.getElementById("toPayFinal");
-    if (localStorage.length != 0) {
-        localStorage.clear();
-        calculateTotalPrice();
-        renderToCard();
+    if (confirm("Drücke ok für Bestellung abschliessen!")) {
+        if (localStorage.length != 0) {
+            localStorage.clear();
+            calculateTotalPrice();
+            renderToCard();
+        }
     }
 }
